Fixes #798
When a user tries to add a 2nd adapter, Lora and AdaLora make some checks to
ensure the new adapter is compatible with existing adapters. Currently, that
check is performed halfway through the method. This means that if the check
fails, the new adapter is partially applied, leaving the model in a bad state.
The main purpose of this PR is to ensure that the model state is correct after
such a failure is encountered.

Tests were added to catch this potential bug.

While working on this, I also did some related, but not strictly necessary
changes to the add_adapter methods:

- So far, PeftModel would initialize the underlying Lora/AdaLora/IA³ model by
  passing the nested dict as config (i.e. config={adapter_name:
  adapter_config}). This has been changed so that the adapter config is passed
  directly (i.e. config=adapter_config) and the Lora/AdaLora/IA³ is now
  responsible for adding the adapter config to the config dict. This makes it
  easier to initialize Lora/AdaLora/IA³ directly and should also address #791.
- add_adapter no longer accepts config=None. AFAICT, this possibility was never
  used. TBH I'm unsure if it even makes sense. Please LMK if it should be added
  back in, and what the purpose is of that option.
- More consistency between Lora/AdaLora/IA³ when adding configs. More code is
  reused between Lora and AdaLora, and IA³ is harmonized.
- Add some type annotations
- Extend docstrings to contain adapter_name

I might be lacking some background info to explain some of the choices that were
made here, so please check if my changes make sense and are not breaking
existing code.