Raise error on wrong type for to modules_to_save (#1496)
Resolves #1492

This PR is for user convenience. When they try to pass the wrong type to
modules_to_save, they will now get an early error message, instead of
getting an obscure error when calling forward later.

Note:

The reason why ModulesToSaveWrapper cannot support ModuleDict et al. is
because it tries to call forward on the original or copied module, but
these modules don't implement a forward method.