[WIP] DoRA: Re-use base result when no dropout
This is an experimental branch with the goal of optimizing DoRA
performance. In case there is no LoRA dropout, it should be possible to
re-use the base result from the base model instead of re-calculating it
for the DoRA output.

For now, this is only implemented for lora.Linear and the
test_disable_adapters_with_merging tests are failing.

This also makes the code quite a bit more complicated but could
potentially save some significant extra compute and memory that would be
needed otherwise, as the base results are relatively big.